get_wait_cor_summary works with the new version of find_correlation. Also add some verbose for debugging.
add some verbose to print the variable with which the correlation coefficient with waiting time  is computed.
add two new arguments to the function get_wait_cor_summary, "Verbose" to activate this "debugging option" and "returnCI" such that the function find_correlation can also return the 0.95 level CI for the correlation coefficient.